requirements.txt
# Core deep learning
# Note: If you need a specific CUDA build of PyTorch, install torch/torchvision
# from the official PyTorch index before or instead of this line.
torch>=2.2.0

# Numeric / image / config utilities
numpy>=1.24.0
Pillow>=10.0.0
PyYAML>=6.0.1
requests>=2.31.0
